requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0

# Data Processing
pandas==2.0.3
numpy==1.24.3

# Visualization
plotly==5.18.0

# HTTP Client (for Hugging Face API)
requests==2.31.0
httpx==0.26.0

# Databricks Integration
databricks-sql-connector==3.0.1

# Delta Lake (for local development)
deltalake==0.14.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# AI Integration
huggingface_hub>=0.20.0

# CORS
python-multipart==0.0.6
